[REFACTOR](FITNESS) - Organização de arquivos e imports
