+++ src/pages/ModoFitness/Dieta.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import SideBar from '../../components/Sidebar/SideBar'
 import './Fitness.css'
-import Dieta from '../../components/ModoFitness/Dieta React'
+import Dieta from '../../components/ModoFitness/Dieta'
 
 const DietaPage = () => {
   return (
